evaluation_system_v2/requirements.txt
# AI Doc Checker - Evaluation System V2 Requirements

# Core ML/AI libraries
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.30.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0

# PDF processing  
PyMuPDF>=1.20.0

# HTTP requests
requests>=2.28.0

# Data handling
numpy>=1.21.0
pandas>=1.4.0

# JSON and file handling (built-in)
# json, os, datetime, base64, io, tempfile, shutil

# Optional: Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=7.6.0

# Optional: Progress bars
tqdm>=4.64.0

# Optional: Logging
loguru>=0.6.0 